Hey,
In Priority, the word high was spelled wrong and I corrected it.

As for my actual task, I have done what I can in teh Assignments Model. I used Emplyee as my orientation. However, I have commented out the fields of Language and Employee (which I added). I think that the connection to the Task and Schedule should be enough, where the Employee is already there.

Also I see two different schemes for the relations between the tables. I was not sure if I should use the old one or the latest one posted by shady so I used the latest one and used ManyToOne. I will push my commit for now and if there is anything else please tell me and we can revert my commit.

NOTE: I seem to have an error annotation is not allowed here when I try to write the ManyToOne Relation. I do not get that error anywhere else except on what I am working on (Assignments model). I have checked online and people keep referring back to Spring methods which need implementations that I do not really understand. I am assuming that my Intellij is working up and not making sense since it has already done so in the past. Tell me If you see the same thing or if you know what exactly is wrong.

Simply send me on slack and I will answer as soon as possible. i will be online on slack today as well.